# lib/api-zod/src/generated/types/radarMonitorRun.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { RadarMonitorRunStatus } from './radarMonitorRunStatus';
import type { RadarMonitorRunTrigger } from './radarMonitorRunTrigger';

export interface RadarMonitorRun {
  id: string;
  source_id: string;
  source_label: string;
  trigger: RadarMonitorRunTrigger;
  status: RadarMonitorRunStatus;
  started_at: string;
  finished_at: string | null;
  /** @minimum 0 */
  attempts: number;
  /** @minimum 0 */
  item_count: number;
  /** @minimum 0 */
  change_count: number;
  http_status: number | null;
  error_message: string;
  duration_ms: number | null;
}