# lib/api-zod/src/generated/types/radarMonitorRunResult.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { RadarChangeEvent } from './radarChangeEvent';
import type { RadarMonitorRun } from './radarMonitorRun';

export interface RadarMonitorRunResult {
  runs: RadarMonitorRun[];
  changes: RadarChangeEvent[];
}