# lib/api-zod/src/generated/types/radarMonitorSourceStatus.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { RadarMonitorSourceStatusConnector } from './radarMonitorSourceStatusConnector';
import type { RadarMonitorSourceStatusLastStatus } from './radarMonitorSourceStatusLastStatus';

export interface RadarMonitorSourceStatus {
  source_id: string;
  source_label: string;
  connector: RadarMonitorSourceStatusConnector;
  endpoint_url: string;
  enabled: boolean;
  last_status: RadarMonitorSourceStatusLastStatus;
  last_run_at: string | null;
  next_run_at: string | null;
  last_error: string;
  /** @minimum 0 */
  consecutive_failures: number;
}