# lib/api-zod/src/generated/types/radarMonitorStatus.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { RadarChangeEvent } from './radarChangeEvent';
import type { RadarMonitorRun } from './radarMonitorRun';
import type { RadarMonitorSourceStatus } from './radarMonitorSourceStatus';
import type { RadarMonitorStatusSummary } from './radarMonitorStatusSummary';

export interface RadarMonitorStatus {
  summary: RadarMonitorStatusSummary;
  sources: RadarMonitorSourceStatus[];
  recent_runs: RadarMonitorRun[];
  recent_changes: RadarChangeEvent[];
}