# lib/api-zod/src/generated/types/radarSourceInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { RadarSourceInputConnector } from './radarSourceInputConnector';

export interface RadarSourceInput {
  id?: string;
  termino: string;
  tipo: string;
  frecuencia: string;
  notas: string;
  connector?: RadarSourceInputConnector;
  endpoint_url?: string;
  enabled?: boolean;
  competitor_id?: string | null;
  [key: string]: unknown;
 }